LPD-57377 This belongs to batch engine
