+++ modules/apps/batch-engine/batch-engine-api/bnd.bnd
@@ -16,6 +16,7 @@ Export-Package:\
 	com.liferay.batch.engine.service,\
 	com.liferay.batch.engine.service.persistence,\
 	com.liferay.batch.engine.strategy,\
+	com.liferay.batch.engine.thread.local,\
 	com.liferay.batch.engine.unit,\
 	com.liferay.batch.engine.zip
 Import-Package:\
