+++ modules/apps/batch-engine/batch-engine-api/src/main/resources/com/liferay/batch/engine/thread/local/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
